// src/models/PublicProfile.ts
export interface PublicProfile {
  email: string;
  languages: any[];
  educations: any[];
  skills: any[];
  timeline: any[];
  fullName: string;
  location: string;
  personalStatement: string;
  phoneNumber: string;
  position: string;
  showLevel: boolean;
  displayName: string;
  profilePicture: string;
  projects: any[];
  isFollowed: boolean;
  totalFollower: number;
  totalFollowing: number;
  page: string;
  id: string;
}